+++ assets/css/main.css
@@ -77,45 +77,40 @@
 .header__button .header__button-reservar:hover {
   background-color: rgb(0, 98.4, 204);
 }
-.header #banner .contenedor {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-}
-.header #banner img {
-  background-position: center;
+.header #banner .banner {
+  background-image: url("../../img/banner.jpg");
   background-size: cover;
-}
-.header #banner h1, .header #banner .h1 {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -100%);
-  font-size: 6rem;
+  background-position: center;
+  width: 100%;
+  height: 100vh; /* 100% de la altura de la ventana */
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
   text-align: center;
-  padding: 10px;
-  width: 50%;
-  font-weight: bold;
-  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
-  color: #007bff;
+  color: white;
 }
-.header #banner p {
-  width: 50%;
-  padding: 15px 0;
+.header #banner .banner h1, .header #banner .banner .h1 {
+  font-size: 3em;
+  margin: 0.5em 0;
+  color: red;
 }
-.header #banner .btn {
-  display: inline-block;
-  font-size: 15px;
-  padding: 12px 23px;
+.header #banner .banner p {
+  font-size: 1.5em;
+  margin: 0.5em 0;
+  Color: red;
+}
+.header #banner .banner button {
+  padding: 1em 2em;
+  font-size: 1em;
+  color: white;
+  background-color: rgba(0, 0, 0, 0.5);
+  border: none;
   cursor: pointer;
-  font-weight: bold;
-  letter-spacing: 1px;
-  border-radius: 4px;
+  transition: background-color 0.3s ease;
 }
-.header #banner .btn-dark {
-  background: primary-color;
-  border: 1px;
-  color: #fff;
+.header #banner .banner button:hover {
+  background-color: rgba(0, 0, 0, 0.8);
 }
 
 .vision-mission {
