+++ contacto.html
@@ -83,6 +83,6 @@
       </div>
   </div>
 </footer>
-<<script src="assets/js/main.js"></script>!-- Fin de footer -->
+<<script src="assets/js/app.js"></script>!-- Fin de footer -->
 </body>
 </html>
