+++ equipo.html
@@ -30,19 +30,12 @@
 <!--Fin de header  -->
 
 <!-- Inicio grid equipo medico-->
-    
-
 <section class="testimonials">
     <div class="container">
-            <h2 class="testimonials__title">Nuestro Equipo Médico</h2>
+        <h2 class="testimonials__title">Nuestro Equipo Médico</h2>
         <div class="row testimonials__grid">
             <div class="col-md-4 testimonials__col">
-                <div class="testimonials__item">
-                    <img src="assets/img/doctor1.jpeg" alt="testimonials__image"> 
-                    <h2 class="testimonials__name">Felipe González</h2>
-                    <h2 class="testimonials__name">Cardiología</h2>
-                    <p class="testimonials__text">Médico Cirujano, Universidad Nacional De Cordoba, Argentina</p>
-                </div>
+               
             </div>
             <div class="col-md-4 testimonials__col">
                 <div class="testimonials__item">
@@ -100,6 +93,7 @@
             </div>
         </div>
     </footer>
+    <script type="module" src="assets/js/app.js"></script>
 <!-- Fin de footer -->
 </body>
 </html>
