+++ index.html
@@ -30,11 +30,11 @@
                     </div>
                 </div>
                 <div id="banner">
-                    <div class="container">
-                        <h1 class="banner__title">Hospital San Clemente</h1>
-                        <img src="assets/img/banner.jpg"alt="banner">
+                    <div class="banner">
+                        <!-- <img src="assets/img/banner.jpg" alt="banner"> -->
+                        <h1>Hospital San Clemente</h1>
                         <p>Gracias por confiar en nosotros y nuestros equipos</p>
-                        <a href="contacto.html" class="btn btn-dark">Contáctanos</a>
+                        <a href="contacto.html" class="btn btn-dark">Contacto</a>
                     </div>
                 </div>
             </div>
